refactor(connifo): create `types.ts` for type definitions
